requirements.txt
google-api-python-client
oauth2client
python-dateutil
requests
flask
flask_cors
lifxlan
flux_led
gevent
spotipy
meross_iot
pytz
gTTS
playsound